Revert "Réglages de bugs"
This reverts commit 5613c128088c43734b42bd741403a31c859a450b.

+++ main.cpp
@@ -152,7 +152,7 @@ int main()
             if (num1 == 2)
             {
 
-                for(int i = 0 ; i < (int)vec_sommets.size() ; ++i)
+                for(int i = 0 ; i < vec_sommets.size() ; ++i)
                 {
                     if(num2 != vec_sommets[i]->getNum())
                     {
